requirement.txt
Flask
Pillow
easyocr
torch
